feat：“项目初始化，完善问题表点赞和收藏，以及一些义务的修改”

+++ src/main/java/com/cool/pandora/model/vo/QuestionViewVO.java
@@ -21,17 +21,17 @@ public class QuestionViewVO implements Serializable {
     /**
      * 浏览记录Id
      */
-    private Integer viewId;
+    private Long viewId;
 
     /**
      * 访问用户Id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 被浏览的面试题的唯一标识
      */
-    private Integer questionId;
+    private Long questionId;
 
     /**
      * 用户访问该题目的时间戳
